Rewrite spech5 module
